requirements.txt
dnspython
ipwhois
tqdm